mmu_enabled realized

+++ mmu.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint8_t mmu_enabled(void);
+uint16_t mmu_enabled(void);
 uint32_t mmu_get_physical(uint16_t vadr);
 void mmu_use_ispace(void);
 void mmu_cycle(void);
